Support initial_delay parameter
This makes the wait-for argument of
JarvusInnovations_background-action configurable.

initial_delay is used because other inputs are underscored, so this
avoids wait-for/wait_for confusion with the action it calls.

The motivation is that my use case comes up in something more like a
second or two. The initial delay was confusing--the workload appeared
just stuck--but the current value is left as the input default.